Move attn processor to pipeline.
