+++ src/diffusers/models/attention.py
@@ -19,7 +19,7 @@ from torch import nn
 
 from ..utils import maybe_allow_in_graph
 from ..utils.import_utils import is_xformers_available
-from .attention_processor import Attention, TuneAVideoAttnProcessor
+from .attention_processor import Attention
 from .cross_attention import CrossAttention
 from .embeddings import CombinedTimestepLabelEmbeddings
 
@@ -400,7 +400,6 @@ class BasicSparseTransformerBlock(nn.Module):
             bias=attention_bias,
             cross_attention_dim=cross_attention_dim if only_cross_attention else None,
             upcast_attention=upcast_attention,
-            processor=TuneAVideoAttnProcessor(),
         )
         self.norm1 = AdaLayerNorm(dim, num_embeds_ada_norm) if self.use_ada_layer_norm else nn.LayerNorm(dim)
