+++ src/diffusers/pipelines/tune_a_video/pipeline_tune_a_video.py
@@ -5,6 +5,7 @@ from typing import Callable, List, Optional, Union
 
 import numpy as np
 import torch
+import torch.nn.functional as F
 from transformers import CLIPTextModel, CLIPTokenizer
 
 from ...loaders import TextualInversionLoaderMixin
@@ -72,6 +73,62 @@ def tensor2vid(video: torch.Tensor) -> List[np.ndarray]:
     return images
 
 
+class TuneAVideoAttnProcessor:
+    def __call__(self, attn, hidden_states, encoder_hidden_states=None, attention_mask=None, video_length=None):
+        batch_size, sequence_length, _ = hidden_states.shape
+
+        encoder_hidden_states = encoder_hidden_states
+
+        if attn.group_norm is not None:
+            hidden_states = attn.group_norm(hidden_states.transpose(1, 2)).transpose(1, 2)
+
+        query = attn.to_q(hidden_states)
+        query.shape[-1]
+        query = attn.head_to_batch_dim(query)
+
+        if attn.added_kv_proj_dim is not None:
+            raise NotImplementedError
+
+        encoder_hidden_states = encoder_hidden_states if encoder_hidden_states is not None else hidden_states
+        key = attn.to_k(encoder_hidden_states)
+        value = attn.to_v(encoder_hidden_states)
+
+        former_frame_index = torch.arange(video_length) - 1
+        former_frame_index[0] = 0
+
+        # key = rearrange(key, "(b f) d c -> b f d c", f=video_length)
+        key = key.reshape([-1, video_length, *key.shape[1:]])
+        key = torch.cat([key[:, [0] * video_length], key[:, former_frame_index]], dim=2)
+        # key = rearrange(key, "b f d c -> (b f) d c")
+        key = key.flatten(0, 1)
+
+        # value = rearrange(value, "(b f) d c -> b f d c", f=video_length)
+        value = value.reshape([-1, video_length, *value.shape[1:]])
+        value = torch.cat([value[:, [0] * video_length], value[:, former_frame_index]], dim=2)
+        # value = rearrange(value, "b f d c -> (b f) d c")
+        value = value.flatten(0, 1)
+
+        key = attn.head_to_batch_dim(key)
+        value = attn.head_to_batch_dim(value)
+
+        if attention_mask is not None:
+            if attention_mask.shape[-1] != query.shape[1]:
+                target_length = query.shape[1]
+                attention_mask = F.pad(attention_mask, (0, target_length), value=0.0)
+                attention_mask = attention_mask.repeat_interleave(self.heads, dim=0)
+
+        attention_probs = attn.get_attention_scores(query, key, attention_mask)
+        hidden_states = torch.bmm(attention_probs, value)
+        hidden_states = attn.batch_to_head_dim(hidden_states)
+
+        # linear proj
+        hidden_states = attn.to_out[0](hidden_states)
+
+        # dropout
+        hidden_states = attn.to_out[1](hidden_states)
+        return hidden_states
+
+
 class TuneAVideoPipeline(DiffusionPipeline, TextualInversionLoaderMixin):
     r"""
     Pipeline for text-to-video generation.
@@ -120,6 +177,7 @@ class TuneAVideoPipeline(DiffusionPipeline, TextualInversionLoaderMixin):
             unet=unet,
             scheduler=scheduler,
         )
+        self.unet.set_attn_processor(TuneAVideoAttnProcessor())
         self.vae_scale_factor = 2 ** (len(self.vae.config.block_out_channels) - 1)
 
     # Copied from diffusers.pipelines.stable_diffusion.pipeline_stable_diffusion.StableDiffusionPipeline.enable_vae_slicing
